[refactor] 파일이름 index로 변경 및 그에따른 리팩토링
