+++ sidebar/src/components/UI/molecules/bucket_create_button/index.js
@@ -1,4 +1,4 @@
-import React from 'react';
+import React, { useEffect } from 'react';
 import { useHistory } from 'react-router-dom';
 import { connect } from 'react-redux';
 import StyledButton from '../../atoms/styled_button/StyledButton';
@@ -18,9 +18,9 @@ const BucketCreateButton = ({ storeTitle, storeDescription, storeDetails }) => {
   const disabled = storeTitle.length * storeDetails.length === 0;
   console.log('disabled:', disabled);
 
-  const onClickHandler = async () => {
+  const onClickHandler = () => {
     const res = await createBucket(storeTitle, storeDescription, storeDetails);
-    history.push({ pathname: '/' });
+    history.replace('/');
   };
 
   return (
