+++ sidebar/src/components/UI/organisms/Buttons/Buttons.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import styled from 'styled-components';
-import BucketCreateButton from '../../molecules/bucket_create_button/BucketCreateButton';
+import BucketCreateButton from '../../molecules/bucket_create_button';
 import ResetButton from '../../molecules/reset_button/ResetButton';
 
 const ButtonsWrapper = styled.div`
